move version to pkg directory
